Created script to automatically create contests
We are also including some templates files. The templates are a latex file
that allows an automated creation of all latex files needed to the contests
PDF. We also added the problem template. This template is used to build the
problem zip file, that later is uploaded to the BOCA system.

The script itself currently have the following features:
 - Creation of latex files to easily generate the contest's PDF.
 - Temporary creation of problems zip files.
 - Upload of problems directly to the BOCA server.
 - Creation of basic users (judges, staff and score).
 - Configuration file allows specification of:
   - Contest information.
   - Problems list.
   - Problems colors.
   - Custom problems timelimits (override problem's default)

Signed-off-by: Vinicius Kwiecien Ruoso <vinicius.ruoso@gmail.com>